requirements.txt
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.12.4
pytz==2022.1
psycopg2-binary==2.8.6
pytest==7.1.2
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-django==4.5.2
pytest-drf==1.1.2
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
